correct bf16

+++ ch3/MultiHeadAttention.py
@@ -1,3 +1,4 @@
+import logging
 import math
 
 import einops
@@ -110,6 +111,6 @@ def scaled_dot_product_attention(Q: torch.Tensor, K: torch.Tensor, V: torch.Tens
     scaled_qkt = qkt / math.sqrt(d_k)
     if mask is not None:
         scaled_qkt = scaled_qkt.masked_fill(~mask, float("-inf"))
-    softmax_qk = softmax(scaled_qkt, scaled_qkt.dim() - 1)
+    softmax_qk = softmax(scaled_qkt, scaled_qkt.dim() - 1).to(dtype=V.dtype)
     result = einops.einsum(softmax_qk, V, "... seq_len_q seq_len_k, ... seq_len_k d_v -> ... seq_len_q d_v")
     return result
